Update core.asm
In a Ghost battle, the Pokémon behind the Ghost will be identified as seen in the Pokédex, 'hinting' at what Pokémon species is behind the Ghost. This fix checks if you are in a Ghost battle and if so, skips over adding the Pokémon species to the Pokédex as seen.